[android] Move dismiss/postpone receivers to SchedulerService and update notificaiton request codes
